feat: root path에 esbuild 설정 추가
