+++ packages/esbuild-config/README.md
@@ -1,0 +1,1 @@
+# esbuild-config
